refactoring 2_0

+++ app/src/main/java/com/example/pokedex_2_0/data/models/PokemonUiInfoEntity.kt
@@ -6,8 +6,8 @@ import com.example.pokedex_2_0.data.models.request.pokemondetail.Type
 data class PokemonUiInfoEntity(
     val id: Int = -1,
     val name: String = "",
-    val height: Int = -1,
-    val weight: Int = -1,
+    val height: Float = -1f,
+    val weight: Float = -1f,
     val stats: List<Stat> = emptyList(),
     val types: List<Type> = emptyList()
 )
